requirements.txt
Django
djongo
djangorestframework<4
py
pytest